IoTWebServer/requirements.txt
Flask==1.1.2
google-cloud-datastore==2.0.1
google-cloud-firestore==2.0.1
google-cloud-storage==1.33.0
google-cloud-logging==1.15.1
google-cloud-error-reporting==1.0.0
sendgrid
gunicorn==20.0.4
six==1.15.0
google-auth==1.23.0
requests==2.25.0
